+++ tests/test_events.rs
@@ -576,12 +576,11 @@ static mut HOOK_COUNT: u64 = 0;
 
 #[allow(clippy::trivially_copy_pass_by_ref)]
 fn count_event_hook(
-  widget: &mut dyn Widget<Event>,
+  widget: &dyn Widget<Event>,
   _cap: &dyn Cap,
   _event: &Event,
 ) -> Option<UiEvents> {
   assert!(widget.downcast_ref::<TestWidget>().is_some());
-  assert!(widget.downcast_mut::<TestWidget>().is_some());
 
   unsafe {
     HOOK_COUNT += 1;
@@ -649,7 +648,7 @@ fn hook_events_handler() {
 
 
 #[allow(clippy::trivially_copy_pass_by_ref)]
-fn quit_event_hook(_: &mut dyn Widget<Event>, _cap: &dyn Cap, _event: &Event) -> Option<UiEvents> {
+fn quit_event_hook(_: &dyn Widget<Event>, _cap: &dyn Cap, _event: &Event) -> Option<UiEvents> {
   Some(UiEvent::Quit.into())
 }
 
@@ -690,11 +689,7 @@ fn hook_events_with_return() {
 
 
 #[allow(clippy::trivially_copy_pass_by_ref)]
-fn emitting_event_hook(
-  _: &mut dyn Widget<Event>,
-  _cap: &dyn Cap,
-  event: &Event,
-) -> Option<UiEvents> {
+fn emitting_event_hook(_: &dyn Widget<Event>, _cap: &dyn Cap, event: &Event) -> Option<UiEvents> {
   assert_eq!(*event, Event::Key('y'));
 
   Some(Event::Key('z').into())